+++ client/src/planning/ExportUtil.ts
@@ -8,10 +8,10 @@ import {
     Uri, window, workspace, Range, ViewColumn
 } from 'vscode';
 
-import { existsSync } from 'fs'
+import { exists } from '../utils';
 
 export async function exportToAndShow(text: string, uri: Uri): Promise<boolean> {
-    let fileExists = await existsSync(uri.fsPath);
+    let fileExists = await exists(uri.fsPath);
     if (!fileExists) {
         uri = uri.with({ scheme: 'untitled' });
     }
@@ -19,5 +19,10 @@ export async function exportToAndShow(text: string, uri: Uri): Promise<boolean> 
     let newDocument = await workspace.openTextDocument(uri);
     let editor = await window.showTextDocument(newDocument, { viewColumn: ViewColumn.Active, preserveFocus: true });
     let fullRange = newDocument.validateRange(new Range(0, 0, newDocument.lineCount, 0));
-    return editor.edit(edit => edit.replace(fullRange, text));
+    if (await editor.edit(edit => edit.replace(fullRange, text))) {
+        return newDocument.save();
+    }
+    else {
+        throw new Error("Cannot insert text to document: " + newDocument.fileName);
+    }
 }